🧩 Refactor(DevicesNetwork): Move UDP parts to `DevicesDiscoveryServer.cs` and `NetworkHelper.cs`.
